Revert "rdma-cm related fix for HA. sceanrio:
connect req is recieved from the client, but there is no listener - so the rdma cm rejects the connection. However, the reject failed because there was no GRH flag and header set. solution: from the poll_cq we set the flag and from process mad the sgid and dgid are set\n"

this fix showed heavy regresion on good path
This reverts commit b2b4557541913074553b2a8933870733e69076f0.