Submitting dma copy request in case of error ----------------------------------------------
In the case an error occurres after pushing some requests to the dma
engine a submit call must be performed. If not, the dma engine may
remain in a colatile state - having entries in its ring which are not to
be processed. This commit ensures that indeed a submit request is
carried to the engine in the case of such an error