Enable compiling with define_trace.h locally
There are issues with finding local trace header file
as the compiler does not detect the file in "." folder
which is the kernel source root directory.
if trying to compile locally it fails.

This patch fixes it.
the instructions taken from https://lwn.net/Articles/383362/
and probably the best practice for using TRACE_EVENT() macro